+++ types/mikro-orm/index.d.ts
@@ -4,6 +4,12 @@ export class VectorType extends Type<string, string> {
     convertToJSValue(value: any, platform: any): any;
     getColumnType(prop: any, platform: any): any;
 }
+export class HalfvecType extends Type<string, string> {
+    constructor();
+    convertToDatabaseValue(value: any, platform: any): string;
+    convertToJSValue(value: any, platform: any): any;
+    getColumnType(prop: any, platform: any): any;
+}
 export function l2Distance(column: any, value: any, em: any): any;
 export function maxInnerProduct(column: any, value: any, em: any): any;
 export function cosineDistance(column: any, value: any, em: any): any;